Use cleanupcode

+++ Client/Services/Synchronizer/Synchronizer.cs
@@ -1,4 +1,3 @@
-using System.Net.Http.Json;
 using Couple.Client.Infrastructure;
 using Couple.Client.Model.Calendar;
 using Couple.Client.Model.Issue;
@@ -6,17 +5,18 @@ using Couple.Client.States.Calendar;
 using Couple.Client.States.Issue;
 using Couple.Shared.Model.Change;
 using Microsoft.JSInterop;
+using System.Net.Http.Json;
 
 namespace Couple.Client.Services.Synchronizer;
 
 public class Synchronizer
 {
-    private readonly IJSRuntime _js;
+    private readonly EventStateContainer _eventStateContainer;
 
     private readonly HttpClient _httpClient;
 
     private readonly IssueStateContainer _issueStateContainer;
-    private readonly EventStateContainer _eventStateContainer;
+    private readonly IJSRuntime _js;
 
     public Synchronizer(IJSRuntime js,
         HttpClient httpClient,
@@ -44,10 +44,7 @@ public class Synchronizer
             .Select(change => change.Id)
             .ToList());
 
-        if (idsToDelete.Guids.Any())
-        {
-            await _httpClient.DeleteAsJsonAsync("api/Changes", idsToDelete);
-        }
+        if (idsToDelete.Guids.Any()) await _httpClient.DeleteAsJsonAsync("api/Changes", idsToDelete);
 
         var issues = _js.InvokeAsync<List<IssueModel>>("getIssues").AsTask();
         var eventsTask = _js.InvokeAsync<List<EventModel>>("getAllEvents").AsTask();
